+++ kmcp/cmd/util-hash.go
@@ -54,24 +54,6 @@ func CalcFPR(numElements uint64, numHashes int, signatureSize uint64) float64 {
 	return math.Pow(1-math.Pow(math.E, float64(-numHashes)*float64(numElements)/float64(signatureSize)), float64(numHashes))
 }
 
-/*
-p, fpr of single bloom filter.
-k, theshold of query coverage.
-l, number of k-mers.
-
-import math
-fpr = lambda p,k,l: math.exp(-l * (k - p) * (k - p) / (2 * (1 - p))
-
-fpr(0.3, 0.8, 60)
-*/
-func maxFPR(p float64, k float64, l int) float64 {
-	return math.Exp(-float64(l) * (k - p) * (k - p) / (2 * (1 - p)))
-}
-
-func maxFPRf(p float64, k float64, l float64) float64 {
-	return math.Exp(-l * (k - p) * (k - p) / (2 * (1 - p)))
-}
-
 // get the two basic hash function values for data.
 // Based on early version of https://github.com/willf/bloom/blob/master/bloom.go .
 func baseHashes(hash uint64) (uint32, uint32) {